Connector : * Merge files * delete already merged file
